API/main-api/requirements.txt
fastapi
fastapi-keycloak
uvicorn
python-keycloak
pydantic_settings
python-multipart
slowapi 
email-validator
psycopg2